third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/script-formatter-breakpoints-2-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointsSetAndRemoveInFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed
unformatted2.js:formatted:4checked    var c = 3;
Breakpoint sidebar pane while paused in raw
No breakpoints
Script execution resumed.
